// src/models/fixture-status.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GoMoney Basic Football-League Manager Service
 * This service provides endpoints for all `basic football-league manager` related interactions
 *
 * OpenAPI spec version: 1.0.0
 * Contact: fatai@mail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * Possible States or Statuses of a Fixture
 * @export
 * @enum {string}
 */
export enum FixtureStatus {
    PENDING = 'PENDING',
    COMPLETED = 'COMPLETED'
}