// src/models/login-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GoMoney Basic Football-League Manager Service
 * This service provides endpoints for all `basic football-league manager` related interactions
 *
 * OpenAPI spec version: 1.0.0
 * Contact: fatai@mail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { User } from './user';
/**
 * 
 * @export
 * @interface LoginResponse
 */
export interface LoginResponse {
    /**
     * 
     * @type {User}
     * @memberof LoginResponse
     */
    user: User;
    /**
     * User authentication token (JWT)
     * @type {string}
     * @memberof LoginResponse
     */
    token: string;
}